# openapi.yaml
openapi: 3.0.3
info:
  title: Flight Data API
  description: |
    REST API for accessing and managing simulated airline flight data.
    Flight data is auto-generated daily at 00:05 UTC (configurable via `GENERATE_HOUR` / `GENERATE_MINUTE` env vars).
    
    ## Authentication
    
    **Public API** endpoints require a valid API key passed via the `x-api-key` header.  
    **Admin API** endpoints require an authenticated admin session (cookie-based).
  version: 1.0.0
  contact:
    name: Flight Data API Support
servers:
  - url: /
    description: Current host
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key prefixed with `fda_`. Obtain from admin panel.
    AdminSession:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after successful admin login.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
    Flight:
      type: object
      properties:
        id:
          type: integer
          description: Auto-increment primary key
          example: 1
        callsign:
          type: string
          description: Flight callsign
          example: PVL910
        actype:
          type: string
          description: ICAO aircraft type designator
          example: B77E
        acregistration:
          type: string
          description: Aircraft registration
          example: C-GMXE
        departureicao:
          type: string
          description: Departure airport ICAO code
          example: CYUL
        arrivalicao:
          type: string
          description: Arrival airport ICAO code
          example: CYQB
        alternateicao:
          type: string
          description: Alternate/diversion airport ICAO code
          example: CYMX
        departuretime:
          type: string
          format: date-time
          description: Scheduled departure time (UTC)
          example: "2026-03-04T08:15:00.000Z"
        arrivaltime:
          type: string
          format: date-time
          description: Scheduled arrival time (UTC)
          example: "2026-03-04T09:45:00.000Z"
        eta:
          type: string
          format: date-time
          description: Estimated time of arrival (UTC)
          example: "2026-03-04T09:52:00.000Z"
        alteta:
          type: string
          format: date-time
          description: ETA to alternate airport (UTC)
          example: "2026-03-04T09:55:00.000Z"
        generation_date:
          type: string
          format: date
          description: The date this record was generated for
          example: "2026-03-04"
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp
          example: "2026-03-04T00:05:00.000Z"
        updated_at:
          type: string
          format: date-time
          description: Record last-update timestamp
          example: "2026-03-04T00:05:00.000Z"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 252
        pages:
          type: integer
          example: 6
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Description of the error
...
  parameters:
    FlightId:
      name: id
      in: path
      required: true
      description: Flight ID
      schema:
        type: integer
...
paths:
...
  /api/v1/flights:
    get:
      tags: [Flights]
      summary: List flights
      description: Retrieve paginated flight data with optional filters.
      operationId: getFlights
      security:
        - ApiKeyAuth: []
      parameters:
        - name: date
          in: query
          description: Filter by generation date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: from_date
          in: query
          description: Start of date range (inclusive)
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          description: End of date range (inclusive)
          schema:
            type: string
            format: date
        - name: callsign
          in: query
          description: Exact callsign match
          schema:
            type: string
          example: PVL910
        - name: departure
          in: query
          description: Departure ICAO code
          schema:
            type: string
          example: CYUL
        - name: arrival
          in: query
          description: Arrival ICAO code
          schema:
            type: string
          example: CYQB
        - name: registration
          in: query
          description: Aircraft registration
          schema:
            type: string
          example: C-GMXE
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Results per page (max 200)
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Paginated list of flights
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Flight"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          description: Missing x-api-key header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/flights/{id}:
    get:
      tags: [Flights]
      summary: Get flight by ID
      description: Retrieve a single flight record by its ID.
      operationId: getFlightById
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FlightId"
      responses:
        "200":
          description: Flight details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Flight"
        "401":
          description: Missing x-api-key header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Flight not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ========================================================================
  # PUBLIC API — Dates & Stats
  # ========================================================================
  /api/v1/dates:
    get:
      tags: [Statistics]
      summary: List generation dates
      description: List all dates for which flight data has been generated.
      operationId: getDates
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: List of generation dates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        generation_date:
                          type: string
                          format: date
                          example: "2026-03-04"
                        flights_count:
                          type: integer
                          example: 36
                        created_at:
                          type: string
                          format: date-time
                          example: "2026-03-04T00:05:00.000Z"
        "401":
          description: Missing x-api-key header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/stats:
    get:
      tags: [Statistics]
      summary: Get summary statistics
      description: Returns total flights, total days of data, and the latest generation date.
      operationId: getStats
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      total_flights:
                        type: integer
                        example: 252
                      total_days:
                        type: integer
                        example: 7
                      latest_generation_date:
                        type: string
                        format: date
                        nullable: true
                        example: "2026-03-05"
        "401":
          description: Missing x-api-key header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ========================================================================
  # ADMIN API — Auth
  # ========================================================================
...